# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Realty Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PropertyType = typeof PropertyType[keyof typeof PropertyType];


export const PropertyType = {
  sale: 'sale',
  rent: 'rent',
} as const;

export type PropertyCategory = typeof PropertyCategory[keyof typeof PropertyCategory];


export const PropertyCategory = {
  apartment: 'apartment',
  villa: 'villa',
  office: 'office',
  land: 'land',
  commercial: 'commercial',
} as const;

export interface Property {
  id: number;
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  type: PropertyType;
  category: PropertyCategory;
  price: number;
  priceUnit?: string;
  area: number;
  city: string;
  cityAr: string;
  address: string;
  addressAr: string;
  lat: number;
  lng: number;
  bedrooms: number;
  bathrooms: number;
  images: string[];
  featured: boolean;
  amenities: string[];
  amenitiesAr?: string[];
  /** @nullable */
  yearBuilt?: number | null;
  /** @nullable */
  floor?: number | null;
  /** @nullable */
  totalFloors?: number | null;
  /** @nullable */
  parkingSpaces?: number | null;
  agentName?: string;
  agentNameAr?: string;
  agentPhone?: string;
  agentEmail?: string;
  createdAt: string;
}

export interface PropertyList {
  properties: Property[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type PropertyInputType = typeof PropertyInputType[keyof typeof PropertyInputType];


export const PropertyInputType = {
  sale: 'sale',
  rent: 'rent',
} as const;

export type PropertyInputCategory = typeof PropertyInputCategory[keyof typeof PropertyInputCategory];


export const PropertyInputCategory = {
  apartment: 'apartment',
  villa: 'villa',
  office: 'office',
  land: 'land',
  commercial: 'commercial',
} as const;

export interface PropertyInput {
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  type: PropertyInputType;
  category: PropertyInputCategory;
  price: number;
  priceUnit?: string;
  area: number;
  city: string;
  cityAr: string;
  address: string;
  addressAr: string;
  lat: number;
  lng: number;
  bedrooms: number;
  bathrooms: number;
  images: string[];
  featured?: boolean;
  amenities?: string[];
  amenitiesAr?: string[];
  yearBuilt?: number;
  floor?: number;
  totalFloors?: number;
  parkingSpaces?: number;
  agentName?: string;
  agentNameAr?: string;
  agentPhone?: string;
  agentEmail?: string;
}

export interface PropertyUpdate {
  title?: string;
  titleAr?: string;
  description?: string;
  descriptionAr?: string;
  price?: number;
  featured?: boolean;
  images?: string[];
}

export interface FavoriteInput {
  propertyId: number;
}

export interface StatsSummary {
  totalProperties: number;
  forSale: number;
  forRent: number;
  avgPriceSale: number;
  avgPriceRent: number;
  totalCities: number;
  newThisMonth: number;
}

export interface CityStats {
  city: string;
  cityAr: string;
  count: number;
}

export type ListPropertiesParams = {
type?: ListPropertiesType;
category?: ListPropertiesCategory;
city?: string;
minPrice?: number;
maxPrice?: number;
minArea?: number;
maxArea?: number;
bedrooms?: number;
bathrooms?: number;
featured?: boolean;
search?: string;
sortBy?: ListPropertiesSortBy;
page?: number;
limit?: number;
};

export type ListPropertiesType = typeof ListPropertiesType[keyof typeof ListPropertiesType];


export const ListPropertiesType = {
  sale: 'sale',
  rent: 'rent',
  all: 'all',
} as const;

export type ListPropertiesCategory = typeof ListPropertiesCategory[keyof typeof ListPropertiesCategory];


export const ListPropertiesCategory = {
  apartment: 'apartment',
  villa: 'villa',
  office: 'office',
  land: 'land',
  commercial: 'commercial',
  all: 'all',
} as const;

export type ListPropertiesSortBy = typeof ListPropertiesSortBy[keyof typeof ListPropertiesSortBy];


export const ListPropertiesSortBy = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  newest: 'newest',
  oldest: 'oldest',
  area_asc: 'area_asc',
  area_desc: 'area_desc',
} as const;