# lib/api-zod/src/generated/types/cityStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Realty Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface CityStats {
  city: string;
  cityAr: string;
  count: number;
}