# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Realty Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyCategory } from './propertyCategory';
import type { PropertyType } from './propertyType';

export interface Property {
  id: number;
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  type: PropertyType;
  category: PropertyCategory;
  price: number;
  priceUnit?: string;
  area: number;
  city: string;
  cityAr: string;
  address: string;
  addressAr: string;
  lat: number;
  lng: number;
  bedrooms: number;
  bathrooms: number;
  images: string[];
  featured: boolean;
  amenities: string[];
  amenitiesAr?: string[];
  /** @nullable */
  yearBuilt?: number | null;
  /** @nullable */
  floor?: number | null;
  /** @nullable */
  totalFloors?: number | null;
  /** @nullable */
  parkingSpaces?: number | null;
  agentName?: string;
  agentNameAr?: string;
  agentPhone?: string;
  agentEmail?: string;
  createdAt: string;
}