# lib/api-zod/src/generated/types/propertyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Realty Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyInputCategory } from './propertyInputCategory';
import type { PropertyInputType } from './propertyInputType';

export interface PropertyInput {
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  type: PropertyInputType;
  category: PropertyInputCategory;
  price: number;
  priceUnit?: string;
  area: number;
  city: string;
  cityAr: string;
  address: string;
  addressAr: string;
  lat: number;
  lng: number;
  bedrooms: number;
  bathrooms: number;
  images: string[];
  featured?: boolean;
  amenities?: string[];
  amenitiesAr?: string[];
  yearBuilt?: number;
  floor?: number;
  totalFloors?: number;
  parkingSpaces?: number;
  agentName?: string;
  agentNameAr?: string;
  agentPhone?: string;
  agentEmail?: string;
}