# lib/api-zod/src/generated/types/propertyList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Realty Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { Property } from './property';

export interface PropertyList {
  properties: Property[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}