# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Realty Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyUpdate {
  title?: string;
  titleAr?: string;
  description?: string;
  descriptionAr?: string;
  price?: number;
  featured?: boolean;
  images?: string[];
}