# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Realty Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalProperties: number;
  forSale: number;
  forRent: number;
  avgPriceSale: number;
  avgPriceRent: number;
  totalCities: number;
  newThisMonth: number;
}